requirements.txt
aiofiles==25.1.0
Appium-Python-Client==5.2.7
attrs==25.4.0
certifi==2026.2.25
fb-idb==1.1.7
grpclib==0.4.9
h11==0.16.0
h2==4.3.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.11
multidict==6.7.1
outcome==1.3.0.post0
protobuf==7.34.1
PySocks==1.7.1
selenium==4.41.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
treelib==1.8.0
trio==0.33.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.6.3
websocket-client==1.9.0
wheel==0.45.1
wsproto==1.3.2
